Fixed calculation service test

+++ src/main/java/com/playtech/supermarket/util/MonetaryUtils.java
@@ -5,7 +5,7 @@ import javax.money.MonetaryAmount;
 
 public class MonetaryUtils {
 
-    public static MonetaryAmount createMonetaryAmount(long howMuch){
+    public static MonetaryAmount createMonetaryAmount(double howMuch){
         return Monetary.getDefaultAmountFactory().setCurrency("GBP")
                 .setNumber(howMuch).create();
     }
